+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## Changelog
-
+	
+- 0.1.0:
+  - Added utils.createDir
+	- Added utils.removeDir
+	
+- 0.0.4: Bug Fixes	
 - 0.0.1: Initial Release